GP2-3856 - Core components JavaScript build
