+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 
 ### Enhancements
 - GP2-3852 - Include redirect query parameter in verification link
+- GP2-3856 - Core components JavaScript build
 - GP2-3831 - Analytics event on EP download and delete
 - GP2-3834 - Remove duplicate analytics events
 - GP2-3389 - PII and secret check hooks for pre-commit
