--- core/components/react-components/css-stub.js
@@ -1,1 +1,0 @@
-module.exports = {};
