--- core/components/react-components/webpack.config.js
@@ -1,40 +1,0 @@
-const path = require('path')
-
-module.exports = {
-  entry: {
-    reactComponents: './react-components/src/bundle.js',
-  },
-  output: {
-    path: path.resolve(__dirname, '../static/javascript/'),
-    filename: '[name].js',
-    publicPath: '',
-    library: '[name]',
-    libraryExport: 'default',
-    libraryTarget: 'var',
-  },
-  resolve: {
-    extensions: ['.js', '.jsx'],
-  },
-  module: {
-    rules: [
-      {
-        test: /\.jsx?$/,
-        loader: 'babel-loader',
-        exclude: /node_modules/,
-      },
-      {
-        test: /\.css$/,
-        exclude: /node_modules/,
-        use: [
-          'style-loader',
-          {
-            loader: 'css-loader',
-            options: {
-              modules: true,
-            },
-          },
-        ],
-      },
-    ],
-  },
-}
