+++ core/templates/components/cookie_notice.html
@@ -16,7 +16,7 @@
         </div>
       </div>
       <div class="cookie-confirmation">
-        <div class="column-full">
+        <div class="column-full padding-right-30">
           <p>{% trans "You've accepted all cookies.  You can "%}<a class="link" href="{% url 'core:cookie-preferences' %}">{% trans 'change your settings' %}</a> {% trans 'at any time.' %}</p>
         </div>
       </div>
