+++ core/templates/core/base.html
@@ -13,7 +13,7 @@
 {% endblock %}
 
 {% block head_css %}
-    <link href="{% static 'styles.css' %}" media="all" rel="stylesheet" />
+    <link href="{% static 'magna_styles.css' %}" media="all" rel="stylesheet" />
 {% endblock %}
 
 {% block head_js %}
