+++ domestic/templates/domestic/base.html
@@ -47,7 +47,7 @@
             <link href="{% static 'stylesheets/govuk-elements-styles.min.css' %}" media="all" rel="stylesheet" />
             <link href="{% static 'stylesheets/elements-components.min.css' %}" media="all" rel="stylesheet" />
             {{ block.super }}
-            <link href="{% static 'styles/main.css' %}" rel="stylesheet" type="text/css" />
+            <link href="{% static 'loggedout_styles.css' %}" rel="stylesheet" type="text/css" />
         {% endblock %}
 
         <style>
@@ -68,8 +68,7 @@
 
         {% block head_js %}
             <script src="{% static 'javascript/dit.js' %}"></script>
-            <script src="{% static 'javascript/dit.components.cookie-notice.js' %}"></script>
-            <script src="{% static 'javascript/reactComponents.js' %}"></script>
+            <script src="{% static 'loggedout.js' %}"></script>
             <script>
               // Safari saves a snapshot of a page as it was when the user left, so upon going "back" it doesn’t have to reload and instead it loads the saved snapshot state. That has an unwanted side effect that cookie modal is shown when clicking "go back to the page you came from".
               window.onpageshow = function (event) {
@@ -169,11 +168,10 @@
             <script>
                 var element = document.createElement('div');
                 document.body.appendChild(element);
-                reactComponents.CookiesModal({
+                loggedout.CookiesModal({
                     element: element,
                     privacyCookiesUrl: {{privacy_and_cookies_path}},
-                    preferencesUrl: '{% url "core:cookie-preferences" %}',
-                    isOpen: dit.components.cookieNotice.getPreferencesCookie() === null
+                    preferencesUrl: '{% url "core:cookie-preferences" %}'
                 })
             </script>
             {% endif %}
