+++ package.json
@@ -7,9 +7,7 @@
     "test:dev": "jest --watch",
     "watch": "webpack --watch --config react-components/webpack.config.js --mode development",
     "build-dev": "webpack --config react-components/webpack.config.js --mode development",
-    "build": "npm install great-styles; webpack --config react-components/webpack.config.js --mode production",
-    "build-domestic-components-dev": "webpack --config core/components/react-components/webpack.config.js --mode development",
-    "build-domestic-components": "webpack --config core/components/react-components/webpack.config.js --mode production"
+    "build": "npm install great-styles; webpack --config react-components/webpack.config.js --mode production"
   },
   "jest": {
     "globalSetup": "reacy-components/setupJest.js"
