+++ react-components/src/bundle-loggedout.js
@@ -1,0 +1,7 @@
+import '@babel/polyfill'
+
+import CookiesModal from '@src/components/CookiesModal/CookiesModal'
+
+export default {
+  CookiesModal,
+};
