+++ react-components/src/components/CookiesModal/CookiesModal.css
@@ -1,0 +1,61 @@
+.heading {
+  font-size: 24px;
+  margin: 0;
+}
+
+.synopsis {
+  font-size: 19px;
+  margin-bottom: 30px;
+  margin-top: 30px;
+  text-align: left;
+}
+
+.button {
+  cursor: pointer;
+  font-size: 19px;
+  margin-bottom: 0;
+  height: 52px;
+  padding: 0;
+  line-height: 52px;
+  font-weight: normal;
+}
+
+:global .ReactModal__Overlay {
+  z-index: 1000;
+  background-color: rgba(0, 0, 0, 0.5)!important;
+}
+
+:global .ReactModal__Content {
+  background: #ffffff;
+  padding: 30px!important;
+  inset: 50%!important;
+  bottom: auto;
+  right: auto;
+  transform: translate(-50%, -50%);
+  width: 95%!important;
+  max-height: fit-content;
+  max-width: 600px;
+}
+
+.buttonSeperator {
+  display: inline-block;
+  width: 20px;
+}
+
+@media (min-width: 500px) {
+  :global .ReactModal__Content {
+    padding: 52px!important;
+
+  }
+
+  .buttonContainer {
+    display: flex;
+  }
+  .synopsis {
+    margin-top: 25px;
+    margin-bottom: 45px;
+  }
+  .button {
+    width: 50%;
+  }
+}
