+++ react-components/src/components/CookiesModal/css-stub.js
@@ -1,0 +1,1 @@
+module.exports = {};
